增加新的单元测试内容

+++ tests/PVRExampleTests.cs
@@ -6,7 +6,10 @@ namespace PVRTexLib.Tests
         [OneTimeSetUp]
         public void Setup()
         {
-            Directory.Delete("images/generated", true);
+            if (Directory.Exists("images/generated"))
+            {
+                Directory.Delete("images/generated", true);
+            }
             Directory.CreateDirectory("images/generated");
         }
